refactor: add DonationTicketModal to CausesPage
